both buttons refactored to use base button, allowing both to have a pending state
